[FLINK-24058][coordination] Extract TimerService interface
